laurel_sprout: Kang walleye offsets for GPU composition
Window animations now have rounded corners, meaning that
the shader that we use is more expensive than it used to be.

Because of this, we need to adjust the SF offsets, to make sure
that we'll have enough time to finish the GPU composition without
dropping frames.

These offsets as similar to blueline's, adjusted slightly because
we were getting false "frame missed" signals if composition
starts too early.

Change-Id: I8173e12b006bc1cab6b4cb0b3a5db9f9b3489819

Signed-off-by: mdeejay <mdjrussia@gmail.com>